refactor: Refactor ForgotPassword with a container
