+++ client/src/containers/ForgotPasswordContainer/index.ts
@@ -1,0 +1,1 @@
+export { ForgotPasswordContainer } from "./ForgotPassword.container";
