+++ client/src/containers/index.ts
@@ -1,0 +1,1 @@
+export { ForgotPasswordContainer } from "./ForgotPasswordContainer";
